+++ packages/config/src/ios/Scheme.ts
@@ -18,7 +18,11 @@ export function setScheme(
   config: Pick<ExpoConfig, 'scheme' | 'ios'>,
   infoPlist: InfoPlist
 ): InfoPlist {
-  const scheme = [...getScheme(config), ...getScheme(config.ios ?? {})];
+  const scheme = [
+    ...getScheme(config),
+    // @ts-ignore: TODO: ios.scheme is an unreleased -- harder to add to turtle v1.
+    ...getScheme(config.ios ?? {}),
+  ];
   // Add the bundle identifier to the list of schemes for easier Google auth and parity with Turtle v1.
   if (config.ios?.bundleIdentifier) {
     scheme.push(config.ios.bundleIdentifier);