+++ packages/expo-cli/src/commands/eas-build/utils/git.ts
@@ -11,7 +11,7 @@ import {
   gitStatusAsync,
 } from '../../../git';
 import log from '../../../log';
-import prompts from '../../../prompts';
+import prompts, { confirmAsync } from '../../../prompts';
 
 async function ensureGitRepoExistsAsync(): Promise<void> {
   try {
@@ -27,9 +27,7 @@ async function ensureGitRepoExistsAsync(): Promise<void> {
   log(log.chalk.yellow("It looks like you haven't initialized the git repository yet."));
   log(log.chalk.yellow('EAS Build requires you to use a git repository for your project.'));
 
-  const { confirmInit } = await prompts({
-    type: 'confirm',
-    name: 'confirmInit',
+  const confirmInit = await confirmAsync({
     message: `Would you like to run 'git init' in the current directory?`,
   });
   if (!confirmInit) {
@@ -91,9 +89,7 @@ async function reviewAndCommitChangesAsync(
   await gitDiffAsync();
   log.newLine();
 
-  const { confirm } = await prompts({
-    type: 'confirm',
-    name: 'confirm',
+  const confirm = await confirmAsync({
     message: 'Can we commit these changes for you?',
   });