+++ packages/xdl/src/Project.ts
@@ -1,4 +1,5 @@
 import {
+  ExpoAppManifest,
   ExpoConfig,
   Hook,
   HookArguments,
@@ -41,7 +42,7 @@ import urljoin from 'url-join';
 import { promisify } from 'util';
 import uuid from 'uuid';
 
-import * as Analytics from './Analytics';
+import Analytics from './Analytics';
 import * as Android from './Android';
 import ApiV2 from './ApiV2';
 import Config from './Config';
@@ -52,7 +53,7 @@ import { maySkipManifestValidation } from './Env';
 import { ErrorCode } from './ErrorCode';
 import * as Exp from './Exp';
 import logger from './Logger';
-import { Asset, PublicConfig, exportAssetsAsync, publishAssetsAsync } from './ProjectAssets';
+import { Asset, exportAssetsAsync, publishAssetsAsync } from './ProjectAssets';
 import * as ProjectSettings from './ProjectSettings';
 import * as Sentry from './Sentry';
 import * as ThirdParty from './ThirdParty';
@@ -78,7 +79,7 @@ const treekillAsync = promisify<number, string>(treekill);
 const ngrokConnectAsync = promisify(ngrok.connect);
 const ngrokKillAsync = promisify(ngrok.kill);
 
-type SelfHostedIndex = PublicConfig & {
+type SelfHostedIndex = ExpoAppManifest & {
   dependencies: string[];
 };
 
@@ -171,8 +172,13 @@ function _requireFromProject(modulePath: string, projectRoot: string, exp: ExpoC
 }
 
 // TODO: Move to @expo/config
-export async function getSlugAsync(projectRoot: string): Promise<string> {
-  const { exp } = getConfig(projectRoot, { skipSDKVersionRequirement: true });
+export async function getSlugAsync({
+  projectRoot,
+  exp = getConfig(projectRoot, { skipSDKVersionRequirement: true }).exp,
+}: {
+  projectRoot: string;
+  exp?: Pick<ExpoConfig, 'slug'>;
+}): Promise<string> {
   if (exp.slug) {
     return exp.slug;
   }
@@ -194,7 +200,7 @@ export async function getLatestReleaseAsync(
   const api = ApiV2.clientForUser(user);
   const result = await api.postAsync('publish/history', {
     owner: options.owner,
-    slug: await getSlugAsync(projectRoot),
+    slug: await getSlugAsync({ projectRoot }),
     releaseChannel: options.releaseChannel,
     count: 1,
     platform: options.platform,
@@ -205,6 +211,10 @@ export async function getLatestReleaseAsync(
   } else {
     return null;
   }
+}
+
+function isSelfHostedIndex(obj: any): obj is SelfHostedIndex {
+  return !!obj.sdkVersion;
 }
 
 // Takes multiple exported apps in sourceDirs and coalesces them to one app in outputDir
@@ -244,8 +254,9 @@ export async function mergeAppDistributions(
 
     // put index.jsons into memory
     const putJsonInMemory = async (indexPath: string, accumulator: SelfHostedIndex[]) => {
-      const index = (await JsonFile.readAsync(indexPath)) as SelfHostedIndex;
-      if (!index.sdkVersion) {
+      const index = await JsonFile.readAsync(indexPath);
+
+      if (!isSelfHostedIndex(index)) {
         throw new XDLError(
           'INVALID_MANIFEST',
           `Invalid index.json, must specify an sdkVersion at ${indexPath}`
@@ -387,17 +398,11 @@ export async function exportForAppHosting(
   const iosBundle = bundles.ios.code;
   const androidBundle = bundles.android.code;
 
-  const iosBundleHash = crypto
-    .createHash('md5')
-    .update(iosBundle)
-    .digest('hex');
+  const iosBundleHash = crypto.createHash('md5').update(iosBundle).digest('hex');
   const iosBundleUrl = `ios-${iosBundleHash}.js`;
   const iosJsPath = path.join(outputDir, 'bundles', iosBundleUrl);
 
-  const androidBundleHash = crypto
-    .createHash('md5')
-    .update(androidBundle)
-    .digest('hex');
+  const androidBundleHash = crypto.createHash('md5').update(androidBundle).digest('hex');
   const androidBundleUrl = `android-${androidBundleHash}.js`;
   const androidJsPath = path.join(outputDir, 'bundles', androidBundleUrl);
 
@@ -825,7 +830,7 @@ async function _getPublishExpConfigAsync(
   projectRoot: string,
   options: PublishOptions
 ): Promise<{
-  exp: PublicConfig;
+  exp: ExpoAppManifest;
   pkg: PackageJSONConfig;
 }> {
   if (options.releaseChannel != null && typeof options.releaseChannel !== 'string') {
@@ -836,11 +841,11 @@ async function _getPublishExpConfigAsync(
   // Verify that exp/app.json and package.json exist
   const { exp, pkg } = getConfig(projectRoot);
 
-  if (exp.android && exp.android.config) {
+  if (exp.android?.config) {
     delete exp.android.config;
   }
 
-  if (exp.ios && exp.ios.config) {
+  if (exp.ios?.config) {
     delete exp.ios.config;
   }
 
@@ -850,8 +855,14 @@ async function _getPublishExpConfigAsync(
   if (sdkVersion === 'UNVERSIONED' && !maySkipManifestValidation()) {
     throw new XDLError('INVALID_OPTIONS', 'Cannot publish with sdkVersion UNVERSIONED.');
   }
-  exp.locales = await ExponentTools.getResolvedLocalesAsync(exp);
-  return { exp: { ...exp, sdkVersion: sdkVersion! }, pkg };
+  exp.locales = await ExponentTools.getResolvedLocalesAsync(projectRoot, exp);
+  return {
+    exp: {
+      ...exp,
+      sdkVersion: sdkVersion!,
+    },
+    pkg,
+  };
 }
 
 async function buildPublishBundlesAsync(
@@ -861,16 +872,16 @@ async function buildPublishBundlesAsync(
 ) {
   if (!getenv.boolish('EXPO_USE_DEV_SERVER', false)) {
     try {
-      await startReactNativeServerAsync(
+      await startReactNativeServerAsync({
         projectRoot,
-        {
+        options: {
           nonPersistent: true,
           maxWorkers: publishOptions.maxWorkers,
           target: publishOptions.target,
           reset: publishOptions.resetCache,
         },
-        !publishOptions.quiet
-      );
+        verbose: !publishOptions.quiet,
+      });
       return await fetchPublishBundlesAsync(projectRoot);
     } finally {
       await stopReactNativeServerAsync(projectRoot);
@@ -1013,7 +1024,7 @@ async function _handleKernelPublishedAsync({
 }: {
   projectRoot: string;
   user: User;
-  exp: ExpoConfig;
+  exp: ExpoAppManifest;
   url: string;
 }) {
   let kernelBundleUrl = `${Config.api.scheme}://${Config.api.host}`;
@@ -1022,7 +1033,7 @@ async function _handleKernelPublishedAsync({
   }
   kernelBundleUrl = `${kernelBundleUrl}/@${user.username}/${exp.slug}/bundle`;
 
-  if (exp.kernel.androidManifestPath) {
+  if (exp.kernel?.androidManifestPath) {
     const manifest = await ExponentTools.getManifestAsync(url, {
       'Exponent-SDK-Version': exp.sdkVersion,
       'Exponent-Platform': 'android',
@@ -1036,7 +1047,7 @@ async function _handleKernelPublishedAsync({
     );
   }
 
-  if (exp.kernel.iosManifestPath) {
+  if (exp.kernel?.iosManifestPath) {
     const manifest = await ExponentTools.getManifestAsync(url, {
       'Exponent-SDK-Version': exp.sdkVersion,
       'Exponent-Platform': 'ios',
@@ -1440,17 +1451,21 @@ function _handleDeviceLogs(projectRoot: string, deviceId: string, deviceName: st
     );
   }
 }
-export async function startReactNativeServerAsync(
-  projectRoot: string,
-  options: StartOptions = {},
-  verbose: boolean = true
-): Promise<void> {
+export async function startReactNativeServerAsync({
+  projectRoot,
+  options = {},
+  exp = getConfig(projectRoot).exp,
+  verbose = true,
+}: {
+  projectRoot: string;
+  options: StartOptions;
+  exp?: ExpoConfig;
+  verbose?: boolean;
+}): Promise<void> {
   _assertValidProjectRoot(projectRoot);
   await stopReactNativeServerAsync(projectRoot);
   await Watchman.addToPathAsync(); // Attempt to fix watchman if it's hanging
   await Watchman.unblockAndGetVersionAsync(projectRoot);
-
-  const { exp } = getConfig(projectRoot);
 
   let packagerPort = await _getFreePortAsync(19001); // Create packager options
 
@@ -1974,11 +1989,10 @@ export async function getUrlAsync(projectRoot: string, options: object = {}): Pr
 
 export async function startAsync(
   projectRoot: string,
-  options: StartOptions = {},
+  { exp = getConfig(projectRoot).exp, ...options }: StartOptions & { exp?: ExpoConfig } = {},
   verbose: boolean = true
 ): Promise<ExpoConfig> {
   _assertValidProjectRoot(projectRoot);
-  const { exp } = getConfig(projectRoot);
   Analytics.logEvent('Start Project', {
     projectRoot,
     developerTool: Config.developerTool,
@@ -1994,7 +2008,7 @@ export async function startAsync(
     DevSession.startSession(projectRoot, exp, 'native');
   } else {
     await startExpoServerAsync(projectRoot);
-    await startReactNativeServerAsync(projectRoot, options, verbose);
+    await startReactNativeServerAsync({ projectRoot, exp, options, verbose });
     DevSession.startSession(projectRoot, exp, 'native');
   }